fix(survey): rename daemon.py to daemon_manager.py to resolve package collision
ROOT CAUSE:
- survey-cli/survey/daemon.py (file, DaemonManager class) added in e826fb2
- survey-cli/survey/daemon/ (package, SurveyDaemon + 11 modules) added later in 3eb709c
- Python package always shadows same-named file → 'import survey.daemon' resolves
  to package, making DaemonManager UNREACHABLE.
- survey-cli/survey.py:724 imports 'from survey.daemon import DaemonManager' → fails.
- pkgutil.iter_modules(survey) errors with 'module survey has no attribute daemon'.

FIX:
- Rename daemon.py → daemon_manager.py (preserves all 460 lines unchanged)
- Update survey.py:724 import to from survey.daemon_manager
- Update test_daemon_manager.py mock-paths (16 patches) to survey.daemon_manager
- Update docstring CLI-usage hints in the renamed file

IMPACT:
- main CI has been failing 7+ consecutive runs due to this collision
- All open PRs (#153, #154, #156, #158) inherit this breakage
- This is the SINGLE blocking issue for CI green on main

DOES NOT FIX (separate concerns):
- test_captcha_fallback.py 3 assertion failures (SR-151 logic, pre-existing)

This collision is exactly the drift class SR-159 path-authority will catch.
Refs SR-159.
